add security

+++ src/main/java/org/hostel/service/ApartmentService.java
@@ -26,19 +26,23 @@ public class ApartmentService {
     private final CategoryRepository categoryRepository;
     private final GuestRepository guestRepository;
 
-    public ResponseEntity<?> add(ApartmentDto apartmentDto) {
+    public ResponseEntity<ApartmentDto> add(ApartmentDto apartmentDto) {
         apartmentRepository.save(new Apartment(apartmentDto));
-        return new ResponseEntity<>(HttpStatus.CREATED);
+        return new ResponseEntity<>(apartmentDto, HttpStatus.CREATED);
     }
 
-    public ResponseEntity<?> remove(long id) {
-        apartmentRepository.deleteById(id);
-        return !apartmentRepository.findById(id).isPresent() ? new ResponseEntity<>(HttpStatus.OK) : new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+    public ResponseEntity<ApartmentDto> remove(long id) throws ApartmentNotFoundException {
+        if (apartmentRepository.findById(id).isPresent()) {
+            apartmentRepository.deleteById(id);
+        } else {
+            throw new ApartmentNotFoundException(id);
+        }
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    public ResponseEntity<ApartmentDto> setCategory(long apartmentId, long categoryId) throws ApartmentNotFoundException {
-        Apartment apartment = getById(apartmentId);
-        Category category = categoryRepository.getById(categoryId);
+    public ResponseEntity<ApartmentDto> setCategory(long apartmentId, long categoryId) throws ApartmentNotFoundException, CategoryNotFoundException {
+        Apartment apartment = apartmentRepository.findById(apartmentId).orElseThrow(() -> new ApartmentNotFoundException(apartmentId));
+        Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new CategoryNotFoundException(apartmentId));
         apartment.setCategory(category);
         if (apartment.getCategory().getId() == categoryId) {
             return new ResponseEntity<>(new ApartmentDto(apartment), HttpStatus.OK);
@@ -47,16 +51,15 @@ public class ApartmentService {
     }
 
     public ResponseEntity<List<Guest>> getGuestList(long id) throws ApartmentNotFoundException {
-        List<Guest> guests = guestRepository.findAllByApartment(getById(id));
+        Apartment apartment = apartmentRepository.findById(id).orElseThrow(() -> new ApartmentNotFoundException(id));
+        List<Guest> guests = guestRepository.findAllByApartment(apartment);
         return guests != null && !guests.isEmpty() ? new ResponseEntity<>(guests, HttpStatus.OK) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     public ResponseEntity<Integer> getRoomAmount(long id) throws ApartmentNotFoundException {
-        int roomAmount = getById(id).getRoomAmount();
+        Apartment apartment = apartmentRepository.findById(id).orElseThrow(() -> new ApartmentNotFoundException(id));
+        int roomAmount = apartment.getRoomAmount();
         return roomAmount != 0 ? new ResponseEntity<>(roomAmount, HttpStatus.OK) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
-
-    public Apartment getById(long id) throws ApartmentNotFoundException {
-        return apartmentRepository.findById(id).orElseThrow(() -> new ApartmentNotFoundException(id));
-    }
 }
+
